renamed css folder to scss
