+++ src/components/FeedCard.jsx
@@ -1,7 +1,7 @@
 import Link from 'next/link';
 import React from 'react';
 
-import styles from '../css/card.module.scss';
+import styles from '../scss/card.module.scss';
 
 export default function Card() {
   return (
