+++ src/components/Header.jsx
@@ -2,7 +2,7 @@ import Link from 'next/link';
 import { useRouter } from 'next/router';
 import React, { useState, useContext } from 'react';
 
-import styles from '../css/header.module.scss';
+import styles from '../scss/header.module.scss';
 import * as FirebaseAuth from './FirebaseAuth';
 import DrawerToggleButton from './SideDrawer/DrawerToggleButton';
 import SideDrawer from './SideDrawer/SideDrawer';
