+++ src/components/Settings/BasicInfo.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import styles from '../../css/settings.module.scss';
+import styles from '../../scss/settings.module.scss';
 
 const Basicinfo = () => (
   <div>
