+++ src/components/Settings/SocialHandle.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import styles from '../../css/settings.module.scss';
+import styles from '../../scss/settings.module.scss';
 
 const Social = () => (
   <div>
