+++ src/components/SideDrawer/SideDrawer.jsx
@@ -2,7 +2,7 @@ import Link from 'next/link';
 import PropTypes from 'prop-types';
 import React, { useState } from 'react';
 
-import styles from '../../css/sideDrawer.module.scss';
+import styles from '../../scss/sideDrawer.module.scss';
 import DrawerToggleButton from './DrawerToggleButton';
 
 const SideDrawer = ({ handleClose, router }) => {
