+++ src/scss/projectInfo.module.scss
