+++ src/scss/searchBar.module.scss
